requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
openai==1.3.7
httpx==0.25.2
python-multipart==0.0.6
email-validator==2.1.0.post1
pymongo==4.6.0